pyproject.toml
[project]
name = "smugmug-streamer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=15.0.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "ipdb==0.13.13",
    "python-dotenv>=1.1.1",
    "rauth==0.7.3",
    "requests>=2.32.4",
    "requests-oauthlib>=1.3.1",
    "s3transfer==0.13.1",
]
